+++ core.impl/src/main/java/com/ericsson/bss/cassandra/ecchronos/core/impl/repair/scheduler/ScheduledRepairJobFactory.java
@@ -113,6 +113,7 @@ public final class ScheduledRepairJobFactory
                     .withRepairLockType(myRepairLockType)
                     .withRepairHistory(myRepairHistoryService)
                     .withRepairHistoryProvider(myRepairHistoryService)
+                    .withFaultReporter(myFaultReporter)
                     .build();
         }
         else